fixes applied to live

+++ martini/legacy/scripts/deleteProduct.php
@@ -4,7 +4,8 @@
     $productID = request()->input('productid');	
     $intakeID = request()->input('intakeid');	
     
-    if($productID != ''){
+    if($productID != '' && $productID != null && $productID != -1 && $productID != "-1"){
+        loggedDataChange('product_force_delete',$pallet_id,'User Deleted Product');
 	    deleteProductEntry($productID);
 	    deleteWeightsFor($productID);
     }
